Make .NET 7 isolated the default and only Api project
