+++ Api/WeatherForecastFunction.cs
@@ -1,18 +1,42 @@
 using System;
 using System.Linq;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
+using System.Net;
+using BlazorApp.Shared;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
 
-using BlazorApp.Shared;
-
-namespace BlazorApp.Api
+namespace ApiIsolated
 {
-    public static class WeatherForecastFunction
+    public class HttpTrigger
     {
-        private static string GetSummary(int temp)
+        private readonly ILogger _logger;
+
+        public HttpTrigger(ILoggerFactory loggerFactory)
+        {
+            _logger = loggerFactory.CreateLogger<HttpTrigger>();
+        }
+
+        [Function("WeatherForecast")]
+        public HttpResponseData Run([HttpTrigger(AuthorizationLevel.Anonymous, "get")] HttpRequestData req)
+        {
+            var randomNumber = new Random();
+            var temp = 0;
+
+            var result = Enumerable.Range(1, 5).Select(index => new WeatherForecast
+            {
+                Date = DateTime.Now.AddDays(index),
+                TemperatureC = temp = randomNumber.Next(-20, 55),
+                Summary = GetSummary(temp)
+            }).ToArray();
+
+            var response = req.CreateResponse(HttpStatusCode.OK);
+            response.WriteAsJsonAsync(result);
+
+            return response;
+        }
+
+        private string GetSummary(int temp)
         {
             var summary = "Mild";
 
@@ -26,28 +50,10 @@ namespace BlazorApp.Api
             }
             else if (temp <= 0)
             {
-                summary = "Freezing!";
+                summary = "Freezing";
             }
 
             return summary;
-        }
-
-        [FunctionName("WeatherForecast")]
-        public static IActionResult Run(
-            [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)] HttpRequest req,
-            ILogger log)
-        {
-            var randomNumber = new Random();
-            var temp = 0;
-
-            var result = Enumerable.Range(1, 5).Select(index => new WeatherForecast
-            {
-                Date = DateTime.Now.AddDays(index),
-                TemperatureC = temp = randomNumber.Next(-20, 55),
-                Summary = GetSummary(temp)
-            }).ToArray();
-
-            return new OkObjectResult(result);
         }
     }
 }
