--- ApiIsolated/.vscode/extensions.json
@@ -1,5 +1,0 @@
-{
-    "recommendations": [
-        "ms-azuretools.vscode-azurefunctions"
-    ]
-}
